feat(muta): flush metadata while commit  (#137)
* sort validators

* flush metadata when update_after_commit

* flush metadata when update_after_sync_commit

* flush tieout_gap of mempool every commit

* remove params of config.toml

* uniform type of propose_weight and vote_weight in Validator

+++ core/consensus/src/consensus.rs
@@ -129,8 +129,8 @@ pub fn gen_overlord_status(height: u64, interval: u64, validators: Vec<Validator
         .into_iter()
         .map(|v| Node {
             address:        v.address.as_bytes(),
-            propose_weight: v.propose_weight,
-            vote_weight:    v.vote_weight,
+            propose_weight: v.propose_weight as u8,
+            vote_weight:    v.vote_weight as u8,
         })
         .collect::<Vec<_>>();
